+++ mamar-web/src/app/store/root.ts
@@ -49,7 +49,7 @@ export function rootReducer(root: Root, action: RootAction): Root {
         const newDoc: Doc = {
             id: generateId(),
             bgm: action.bgm ?? new_bgm(),
-            file: action.file,
+            fileHandle: action.file?.handle,
             name: action.name || action.file?.name || "New song",
             isSaved: true,
             activeVariation: 0,